# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Interactive API for testing Neuron keylib functionality",
        "title": "Keylib API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/derive/evm-address": {
            "post": {
                "description": "Derive the Ethereum address from a public key hex string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "derive"
                ],
                "summary": "Derive EVM address from public key",
                "parameters": [
                    {
                        "description": "Public key hex",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.DeriveFromPublicKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.EVMAddressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/derive/evm-address-safe": {
            "post": {
                "description": "Derive the Ethereum address from a public key with explicit error handling for zero keys",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "derive"
                ],
                "summary": "Derive EVM address with error checking",
                "parameters": [
                    {
                        "description": "Public key hex",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.DeriveFromPublicKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.EVMAddressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/derive/peer-id": {
            "post": {
                "description": "Derive the libp2p peer ID from a public key hex string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "derive"
                ],
                "summary": "Derive libp2p peer ID from public key",
                "parameters": [
                    {
                        "description": "Public key hex",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.DeriveFromPublicKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.PeerIDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/derive/public-key": {
            "post": {
                "description": "Derive the public key from a private key hex string",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "derive"
                ],
                "summary": "Derive public key from private key",
                "parameters": [
                    {
                        "description": "Private key hex",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.DeriveFromPrivateKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.PublicKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/encryption/scramble": {
            "post": {
                "description": "Encrypt a private key using AES-256-GCM with Argon2id key derivation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encryption"
                ],
                "summary": "Encrypt a private key",
                "parameters": [
                    {
                        "description": "Private key and password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ScrambleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ScrambleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/encryption/unscramble": {
            "post": {
                "description": "Decrypt an encrypted private key using the password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "encryption"
                ],
                "summary": "Decrypt a private key",
                "parameters": [
                    {
                        "description": "Encrypted key and password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.UnscrambleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.KeyPairResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.DeriveFromPrivateKeyRequest": {
            "type": "object",
            "required": [
                "privateKeyHex"
            ],
            "properties": {
                "privateKeyHex": {
                    "type": "string",
                    "example": "0x1234567890abcdef..."
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.DeriveFromPublicKeyRequest": {
            "type": "object",
            "required": [
                "publicKeyHex"
            ],
            "properties": {
                "publicKeyHex": {
                    "type": "string",
                    "example": "0x02759b048e7ccf6ba68f9658105a4a139b5f9f5dfd451857c600cc28f33a1a99ae"
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.EVMAddressResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "0xe364f2f1e5f4f03d1df682322500b9c68c997ec3"
                },
                "addressChecksum": {
                    "type": "string",
                    "example": "0xe364f2f1e5F4F03d1df682322500b9c68C997ec3"
                },
                "isZero": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.EncryptedKeyDTO": {
            "type": "object",
            "properties": {
                "ciphertext": {
                    "type": "string",
                    "example": "base64..."
                },
                "nonce": {
                    "type": "string",
                    "example": "base64..."
                },
                "salt": {
                    "type": "string",
                    "example": "base64..."
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid hex format"
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.KeyPairResponse": {
            "type": "object",
            "properties": {
                "evmAddress": {
                    "type": "string",
                    "example": "0xe364f2f1e5f4f03d1df682322500b9c68c997ec3"
                },
                "evmAddressChecksum": {
                    "type": "string",
                    "example": "0xe364f2f1e5F4F03d1df682322500b9c68C997ec3"
                },
                "peerID": {
                    "type": "string",
                    "example": "16Uiu2HAm3Lkn9NRieuh3UUTWMNthSDumQL9ctTBKxQqdCC79WUSq"
                },
                "privateKeyHex": {
                    "type": "string",
                    "example": "0x1234567890abcdef..."
                },
                "publicKeyHex": {
                    "type": "string",
                    "example": "0x02759b048e7ccf6ba68f9658105a4a139b5f9f5dfd451857c600cc28f33a1a99ae"
                },
                "publicKeyHexUncompressed": {
                    "type": "string",
                    "example": "0x04759b048e7ccf6ba68f9658105a4a139b5f9f5dfd451857c600cc28f33a1a99ae..."
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.PeerIDResponse": {
            "type": "object",
            "properties": {
                "peerID": {
                    "type": "string",
                    "example": "16Uiu2HAm3Lkn9NRieuh3UUTWMNthSDumQL9ctTBKxQqdCC79WUSq"
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.PublicKeyResponse": {
            "type": "object",
            "properties": {
                "evmAddress": {
                    "type": "string",
                    "example": "0xe364f2f1e5f4f03d1df682322500b9c68c997ec3"
                },
                "evmAddressChecksum": {
                    "type": "string",
                    "example": "0xe364f2f1e5F4F03d1df682322500b9c68C997ec3"
                },
                "peerID": {
                    "type": "string",
                    "example": "16Uiu2HAm3Lkn9NRieuh3UUTWMNthSDumQL9ctTBKxQqdCC79WUSq"
                },
                "publicKeyHex": {
                    "type": "string",
                    "example": "0x02759b048e7ccf6ba68f9658105a4a139b5f9f5dfd451857c600cc28f33a1a99ae"
                },
                "publicKeyHexUncompressed": {
                    "type": "string",
                    "example": "0x04759b048e7ccf6ba68f9658105a4a139b5f9f5dfd451857c600cc28f33a1a99ae..."
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ScrambleRequest": {
            "type": "object",
            "required": [
                "password",
                "privateKeyHex"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "mypassword123"
                },
                "privateKeyHex": {
                    "type": "string",
                    "example": "0x1234567890abcdef..."
                }
            }
        },
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.ScrambleResponse": {
            "type": "object",
            "properties": {
                "encrypted": {
                    "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.EncryptedKeyDTO"
                }
            }
        },
...
        "github_com_aspect-build_neuron-go-hedera-sdk_api_dto.UnscrambleRequest": {
            "type": "object",
            "required": [
                "encrypted",
                "password"
            ],
            "properties": {
                "encrypted": {
                    "$ref": "#/definitions/github_com_aspect-build_neuron-go-hedera-sdk_api_dto.EncryptedKeyDTO"
                },
                "password": {
                    "type": "string",
                    "example": "mypassword123"
                }
            }
        },
...
    }
}